requirements.txt
Django==1.11
djangorestframework==3.6.2
django-allauth==0.30.0
django-rest-auth==0.9.0
django-rest-swagger==2.1.1
django-model-utils==2.5.2
factory-boy==2.8.1
flake8
